+++ internal/daemon/forwarder_watcher_stub.go
@@ -4,8 +4,10 @@ package daemon
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/pidginhost/csm/internal/alert"
+	"github.com/pidginhost/csm/internal/queuehealth"
 )
 
 // ForwarderWatcher is a no-op on non-Linux platforms.
@@ -18,3 +20,5 @@ func NewForwarderWatcher(_ chan<- alert.Finding, _ []string) (*ForwarderWatcher,
 
 // Run is a no-op on non-Linux platforms.
 func (fw *ForwarderWatcher) Run(_ <-chan struct{}) {}
+
+func (fw *ForwarderWatcher) QueueStatuses(time.Time) map[string]queuehealth.Status { return nil }